+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: `SimpleCarousel` Add `id` prop to component (in case of multiple Carousels on one webpage)
+- Fix: Switch `SimpleCarousel` from Radium to Styled Components in order to fix bug for IE11.
 - Fix: IE Flexbox bug on `Header` component CSS
 - Fix: `SearchInput`, remove "clear" cross from IE10+ input.